requirements.txt
ago==0.0.6
alembic==0.8.3
anyjson==0.3.3
blinker==1.4
decorator==4.0.4
Flask==0.10.1
Flask-Cache==0.13.1
Flask-Login==0.2.11
Flask-Mail==0.9.1
Flask-Migrate==1.6.0
Flask-OAuthlib==0.9.2
Flask-Principal==0.4.0
Flask-Script==2.0.5
Flask-SQLAlchemy==2.1
gunicorn==19.3.0
itsdangerous==0.24
Jinja2==2.8
Mako==1.0.3
MarkupSafe==0.23
oauthlib==1.0.3
passlib==1.6.5
pip-review==0.3.7
py-bcrypt==0.4
PyJWT==1.4.0
python-editor==0.4
PyYAML==3.11
requests==2.8.1
requests-oauthlib==0.5.0
SQLAlchemy==1.0.9
Werkzeug==0.11.2
wheel==0.26.0
cryptacular==1.4.1
# Celery
pytz==2015.7
kombu==3.0.32
celery==3.1.19
billiard==3.3.0.22
amqp==1.4.8
redis==2.10.5
paypal==1.2.5
paypalrestsdk==1.11.5
praw==3.3.0
Flask-Admin==1.4.0
rethinkdb==2.3.0.post6
